+++ src/extend-zod.ts
@@ -1,4 +1,5 @@
 import { ZodDate } from "zod";
+import { withGetType } from "zod-to-ts";
 import { ZodDateIn } from "./date-in-schema";
 import { ZodDateOut } from "./date-out-schema";
 import { ZodFile } from "./file-schema";
@@ -13,5 +14,11 @@ export const upload = ZodUpload.create;
  * @deprecated Please use z.dateIn() or z.dateOut() within IO schemas
  * */
 export const date = ZodDate.create;
-export const dateIn = ZodDateIn.create;
-export const dateOut = ZodDateOut.create;
+export const dateIn = (...params: Parameters<typeof ZodDateIn.create>) =>
+  withGetType(ZodDateIn.create(...params), (ts) =>
+    ts.factory.createKeywordTypeNode(ts.SyntaxKind.StringKeyword)
+  );
+export const dateOut = (...params: Parameters<typeof ZodDateOut.create>) =>
+  withGetType(ZodDateOut.create(...params), (ts) =>
+    ts.factory.createKeywordTypeNode(ts.SyntaxKind.StringKeyword)
+  );