+++ src/index.ts
@@ -1,13 +1,8 @@
 export { createConfig, LoggerConfig } from "./config-type";
-export {
-  AbstractEndpoint,
-  EndpointInput,
-  EndpointOutput,
-  EndpointResponse,
-} from "./endpoint";
+export { AbstractEndpoint } from "./endpoint";
 export { Method } from "./method";
 export { EndpointsFactory, defaultEndpointsFactory } from "./endpoints-factory";
-export { IOSchema, FlatObject, markOutput } from "./common-helpers";
+export { IOSchema, FlatObject } from "./common-helpers";
 export { createApiResponse } from "./api-response";
 export { createLogger } from "./logger";
 export { createMiddleware } from "./middleware";
@@ -20,6 +15,7 @@ export { OpenAPI } from "./open-api";
 export { OpenAPIError, DependsOnMethodError, RoutingError } from "./errors";
 export { withMeta } from "./metadata";
 export { testEndpoint } from "./mock";
+export { Client } from "./client";
 
 import * as z from "./extend-zod";
 import createHttpError from "http-errors";